requirements.txt
opencv-python
openvino
onnxruntime
pillow
numpy